Day 4 Part 2
I struggled this for a while since I accidentally had a case set up
where the map keeping track of cards was being initialized incorrectly.
Turns out I didn't even need this initialization logic; by default an
unordered_map of ints will default to 0 if a key is not defined so I
could just use the `+=` operator instead of the bugged initializer.