Day 06 Part 2
Initially I was planning on optimizing the `numberOfWinningRaces()`
method since I observed that the results were symmetrical. But I thought
that was too premature and didn't want to trap myself when Part 2 showed
up. Turns out, that kind of optimization was needed for Part 2!

If I had a nickel for every time I had to typedef something when part 2
comes...